removed some unnecessary diagnostics (CVS internal)

+++ src/ai.hpp
@@ -222,7 +222,7 @@ private:
 
         int best_attack_rating = -1;
         int best_attack = -1;
-        for(int n = 0; n != attacks.size(); ++n) {
+        for(size_t n = 0; n != attacks.size(); ++n) {
 			const battle_stats stats = evaluate_battle_stats(get_info().map,attacker,defender,n,get_info().units,get_info().state,get_info().gameinfo,0,false);
 			const int attack_rating = stats.damage_defender_takes*stats.nattacks*stats.chance_to_hit_defender;
 			if(best_attack == -1 || attack_rating > best_attack_rating) {
